+++ src/shared/components/Drawer/components/DialogSearchAll/services/types/fetchGetSearchAllService.types.ts
