+++ src/shared/components/Drawer/components/DialogSearchAll/types/DialogSearchAll.types.ts
@@ -7,7 +7,7 @@ import {
 import { IDialogSearchAllState } from "shared/common/store/DialogSearchAll/types/DialogSearchAll.types";
 import { IPaginationModel } from "shared/common/types/AppTypes";
 import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
-import { IFetchGetSearchAllResponse } from "shared/components/Drawer/components/DialogSearchAll/service/types";
+import { IFetchGetSearchAllResponse } from "shared/components/Drawer/components/DialogSearchAll/services/types";
 
 export interface IDialogSearchAllForm {
 	query: string;
